+++ phin-vocab/src/main/java/cdc/gov/phinvads/vocab/VocabClient.kt
@@ -32,9 +32,9 @@ class VocabClient {
     fun getValueSetKey(valueSet: ValueSet): StringBuilder {
         val vocabKey = StringBuilder()
         vocabKey.append(valueSet.code)
-        val valueSetVersionResult = service!!.getValueSetVersionByValueSetOidAndVersionNumber(valueSet.oid, 0)
-        val version = valueSetVersionResult.valueSetVersion ?.versionNumber
-        vocabKey.append("_").append("$version")
+      //  val valueSetVersionResult = service!!.getValueSetVersionByValueSetOidAndVersionNumber(valueSet.oid, 0)
+       // val version = valueSetVersionResult.valueSetVersion ?.versionNumber
+       // vocabKey.append("_").append("$version")
        // println("VocabKey :${vocabKey}")
         return vocabKey
     }