+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+### Unreleased
+- Fns-pipeline - first function for POC: Receiver-Debatcher
+
+## [0.0.2] - 2022-09-12 
+ - Developed Receiver-Debatcher function
+ - Developed phin-vads function
+ - First Graphana dashboard with basic infrastructure metrics
+
+commit #96
+
 ## [0.0.1] - 2022-08-29
 
 ### Added
@@ -12,3 +23,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docs: How to create development environment for Azure functions project
 - Tools: Various HL7 development tools or experiments
 - Folders for future tools: mmg, phin-vocab, and fns-hl7-pipeline
+- Phin-vocab function for fetching the value sets from phinvads api.